requirements.txt
requests==2.32.3
pypub3==2.0.8
bs4==0.0.2
tk==0.1.0